storage: Migrate VariantSetStats int fields to long

+++ opencga-storage/opencga-storage-core/src/test/java/org/opencb/opencga/storage/core/variant/VariantStorageTest.java
@@ -33,9 +33,9 @@ public interface VariantStorageTest extends AutoCloseable {
     default void close() throws Exception {}
 
     default int getExpectedNumLoadedVariants(VariantFileMetadata fileMetadata) {
-        int numRecords = fileMetadata.getStats().getVariantCount();
-        return numRecords
-                - fileMetadata.getStats().getTypeCount().getOrDefault(VariantType.SYMBOLIC.name(), 0)
-                - fileMetadata.getStats().getTypeCount().getOrDefault(VariantType.NO_VARIATION.name(), 0);
+        int numRecords = fileMetadata.getStats().getVariantCount().intValue();
+        return (int)(numRecords
+                - fileMetadata.getStats().getTypeCount().getOrDefault(VariantType.SYMBOLIC.name(), 0L)
+                - fileMetadata.getStats().getTypeCount().getOrDefault(VariantType.NO_VARIATION.name(), 0L));
     }
 }
